+++ src/tapis-ui/_wrappers/index.ts
@@ -1,2 +1,3 @@
 export { default as QueryWrapper } from './QueryWrapper';
 export { default as SubmitWrapper } from './SubmitWrapper';
+export { default as Wizard } from './Wizard';
